requirements.txt
# Telegram Bot
python-telegram-bot

# FFmpeg for video processing
moviepy

# YouTube/Google API
google-auth-oauthlib
google-api-python-client

# Environment Variables
python-dotenv

# Scraper (Requests + BeautifulSoup)
requests
beautifulsoup4
